# Little optimization for TakeLast