--- a/src/System.Linq/src/System/Linq/Take.SpeedOpt.cs
+++ b/src/System.Linq/src/System/Linq/Take.SpeedOpt.cs
@@ -30,8 +30,7 @@ namespace System.Linq
                     return partition;
                 }
             }
-
-            if (source is IList<TSource> sourceList)
+            else if (source is IList<TSource> sourceList)
             {
                 if (sourceList.Count > count)
                 {