TakeLast can take advantage of using IPartition or IList directly rather than Enqueue/Dequeue procedure.