@stephentoub 2019-03-14
Cited issues aside (some of which may be contributing to the test failures), @jaredpar and @cston can chime in on whether this is something they'd want to take.  Regardless, thanks for the contribution.