@Romasz 2019-03-15
@stephentoub Thanks for the review. Once and if this one gets accepted I will try to improve SkipLast similar. 

Do you know why UWP is optimized for size rather than speed (even for desktops)? Is there any source I can read something about this?