@stephentoub 2019-03-15
> Do you know why UWP is optimized for size rather than speed (even for desktops)?

Apps are AOT compiled, and the significant generics usage in LINQ leads to massive size increases in some code bases.