@stephentoub 2019-03-15
Can you share benchmarks and resulting performance numbers for all of the various cases covered here, including the cases where optimizations don't end up helping?  There's a cost to this additional interface checking, and in the past it's shown up for small iterator inputs.