@AndyAyersMS 2019-03-15
There are also occasional issues with running out of fast dictionary slots in methods that do a lot of generic type tests -- see eg dotnet/coreclr#22400. I don't think you will hit that here but you might want to verify, and show that when you are testing for multiple types that each extra test is incrementally beneficial to the things you want to speed up and only minimally harmful to the ones you don't.

That is, first measure with no type specialization over your N types. Then add just the first case and remeasure. You should see that one case speed up and all others slow down a tiny bit. Then add the second case; you should see the first case perf remain the same, the second case speed up, and all others slow down a tiny bit more again. Repeat if you have still more types to try.

If you hit dotnet/coreclr#22400 you'll find at some point in this process the previously optimized cases will start to slow down too.