@Romasz 2019-03-15
@stephentoub Can this be a custom project with BenchmarkDotNet? Or it should use the _System.Linq.PerformanceTetsts_ project that's included? The problem with it is that it uses assembly for netcoreapp which doesn't have TakeLast. Is there a configuration for netstandart?