@stephentoub 2019-03-15
> Can this be a custom project with BenchmarkDotNet?

That's fine.

> The problem with it is that it uses assembly for netcoreapp which doesn't have TakeLast

Sorry, I'm not following.  Did you mean it uses netstandard which doesn't have TakeLast?

@adamsitnik, can you help @Romasz with instructions for benchmarking a change he's proposing to a LINQ method?