@Romasz 2019-03-15
> Sorry, I'm not following. Did you mean it uses netstandard which doesn't have TakeLast?

@stephentoub if you take a look at Perf.Linq.cs it doesn't have defined test for TakeLast and also it is not available in assembly used there:

`Assembly netstandard, Version=2.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51 
// C:\Current\GithubProjects\corefx\artifacts\bin\ref\netstandard\netstandard.dll`

@adamsitnik I start building project, any guidance/instructions will be helpful.