@Romasz 2019-03-15
@stephentoub @adamsitnik for start I've made [such project](https://github.com/Romasz/TakeLastPerfTest).

The results are like:

```
BenchmarkDotNet=v0.11.4, OS=Windows 10.0.17763.379 (1809/October2018Update/Redstone5)
Intel Core i7-7820HQ CPU 2.90GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.505
  [Host] : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3324.0
  Core   : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
```


|      Method |  Job | Runtime |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------------ |----- |-------- |-----------:|----------:|----------:|------:|--------:|
| TakeLastOld |  Clr |     Clr | 198.810 us | 0.5743 us | 0.5091 us |  1.00 |    0.00 |
| TakeLastOld | Core |    Core | 199.186 us | 0.4211 us | 0.3733 us |  1.00 |    0.00 |
|             |      |         |            |           |           |       |         |
| TakeLastNew |  Clr |     Clr |  65.061 us | 0.1815 us | 0.1515 us | 33.56 |    0.34 |
| TakeLastNew | Core |    Core |   1.939 us | 0.0184 us | 0.0173 us |  1.00 |    0.00 |