@Romasz 2019-03-16
It turns out that there is a [guide about benchmarking .Net core](https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/benchmarking.md), however it's weird that it's not listed [anywhere here](https://github.com/dotnet/corefx/wiki/New-contributor-Docs#contributing-guide) (unless I've missed it somehow). The link however I have found at [BenchmarkDotNet repo](https://github.com/dotnet/BenchmarkDotNet#our-users). 

Nevertheless, I've managed to build the [benchmarking project](https://github.com/Romasz/CoreFxBenchmarks/tree/master/src/Benchmarks) and use my local build of CoreFx via toolchain. I first build master and benchmarked it and afterwards to the same with the changes. The results are:

**Before the change:**
``` ini

BenchmarkDotNet=v0.11.4, OS=Windows 10.0.17763.379 (1809/October2018Update/Redstone5)
Intel Core i7-7820HQ CPU 2.90GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.505
  [Host]     : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
  Job-MYSMAZ : .NET Core 2c17fabb-c91f-4ea4-ac2c-34544d8b72ce (CoreCLR 4.6.27513.77, CoreFX 4.7.19.16601), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  

```
|             Method |         Mean |      Error |     StdDev |
|------------------- |-------------:|-----------:|-----------:|
|  IEnumManyTakeMany |   884.548 us | 17.6581 us | 17.3426 us |
|  IListManyTakeMany |   234.611 us |  1.1647 us |  1.0895 us |
| CustomManyTakeMany | 1,228.409 us | 17.0377 us | 15.9371 us |
|   IEnumManyTakeFew |   804.662 us | 10.0045 us |  9.3582 us |
|   IListManyTakeFew |   178.044 us |  1.7431 us |  1.6305 us |
|  CustomManyTakeFew | 1,098.817 us | 10.8250 us | 10.1257 us |
|  IEnumLessTakeMany |    12.880 us |  0.0683 us |  0.0639 us |
|  IListLessTakeMany |     7.070 us |  0.0403 us |  0.0377 us |
| CustomLessTakeMany |    15.962 us |  0.1156 us |  0.1081 us |
|    IEnumFewTakeFew |     8.378 us |  0.0345 us |  0.0322 us |
|    IListFewTakeFew |     2.312 us |  0.0139 us |  0.0130 us |
|   CustomFewTakeFew |    11.436 us |  0.1005 us |  0.0940 us |

**After the change:**

``` ini

BenchmarkDotNet=v0.11.4, OS=Windows 10.0.17763.379 (1809/October2018Update/Redstone5)
Intel Core i7-7820HQ CPU 2.90GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.505
  [Host]     : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
  Job-FTITCG : .NET Core 46978f5e-7f77-4598-ba17-462c29c84838 (CoreCLR 4.6.27513.77, CoreFX 4.7.19.16601), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  

```
|             Method |           Mean |         Error |        StdDev |
|------------------- |---------------:|--------------:|--------------:|
|  IEnumManyTakeMany |    55,419.8 ns |    387.431 ns |    362.403 ns |
|  IListManyTakeMany |    12,604.3 ns |     93.503 ns |     82.888 ns |
| CustomManyTakeMany | 1,223,763.6 ns |  9,015.410 ns |  8,433.020 ns |
|   IEnumManyTakeFew |       410.1 ns |      2.340 ns |      2.189 ns |
|   IListManyTakeFew |       180.6 ns |      1.500 ns |      1.404 ns |
|  CustomManyTakeFew | 1,094,212.5 ns | 11,083.325 ns | 10,367.349 ns |
|  IEnumLessTakeMany |     4,928.9 ns |     30.969 ns |     25.860 ns |
|  IListLessTakeMany |     1,068.7 ns |     11.259 ns |     10.532 ns |
| CustomLessTakeMany |    16,160.8 ns |    107.127 ns |    100.207 ns |
|    IEnumFewTakeFew |       402.5 ns |      4.281 ns |      4.004 ns |
|    IListFewTakeFew |       178.3 ns |      1.444 ns |      1.280 ns |
|   CustomFewTakeFew |    11,456.5 ns |     46.616 ns |     38.926 ns |

It took me few minutes to find out that the former is in _us_ and latter in _ns_. 

I haven't managed to run the job on _Crl_ with my local build of _CoreFx._ What I was only able to do is to [create a multi-framework project](https://github.com/Romasz/CoreFxBenchmarks/tree/master/src/TakeLastMethodTests), copy the _TakeLast_ methods as extension ones and benchmark it with _CoreJob_ and _CrlJob_. Results are already [mentioned in the comment above](https://github.com/dotnet/corefx/pull/36051#issuecomment-473445266), more extensive ones are [available at github](https://github.com/Romasz/CoreFxBenchmarks/blob/master/src/TakeLastMethodTests/BenchmarkDotNet.Artifacts/results/TakeLastMethodTests.Program-report-github.md).

A small legend:

-  IEnumManyTakeMany - Enumerable.Range(0, 10000).Select(i => i.ToString()) - take last 1000 items
-  IListManyTakeMany - - Enumerable.Range(0, 10000).Select(i => i.ToString()).ToList() - take last 1000 items
-  CustomManyTakeMany - above enumerator extended by string modification and yielding the result
-  IEnumManyTakeFew - as above but take 3 items
-  IListManyTakeFew - as above but take 3 items
-  CustomManyTakeFew - as above but take 3 items
-  IEnumLessTakeMany - as above but collection has 100 items - get last 1000 items
-  IListLessTakeMany  - as above but collection has 100 items - get last 1000 items
-  CustomLessTakeMany  - as above but collection has 100 items - get last 1000 items
-  IEnumFewTakeFew - as above - collection has 100 items, take last 3
-  IListFewTakeFew - as above - collection has 100 items, take last 3
-  CustomFewTakeFew - as above - collection has 100 items, take last 3